Fine-tune GitHub Actions and versioning (#14)
As it's now, the artifacts from the dev releases overwrite the properly
released ones. Hence, we're switching to using the commit SHA as the
artifact version and release version.

It's also unnecessary and confusing to create proper GitHub releases for
the dev ones, where the point is merely to have the artifacts published
and available for local downstream development.

I don't think we have landed a versioning scheme nor proper release
process, so I'm not sure whether we should stick to using SHA version
for the released versions, but that's more a part of the #13 discussion.

The flatten plugin is required for using this property-based versioning
in a multi-module project. Without it, the version in the subproject
won't be replaced with the property.